requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.0
click==8.1.7
Django==5.0.4
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
psycopg==3.1.12
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.11.0
uritemplate==4.1.1
